+++ python/eggroll/roll_site/conf/mock_roll.json
@@ -4,15 +4,15 @@
     "1"
   ],
   "storage": {
-    "0": "localhost:7778",
-    "1": "localhost:7779"
+    "0": "127.0.0.1:7778",
+    "1": "127.0.0.1:7779"
   },
   "procs": {
     "0": [
-      "localhost:8250"
+      "127.0.0.1:8250"
     ],
     "1": [
-      "localhost:8310"
+      "127.0.0.1:8310"
     ]
   }
 }
