+++ python/eggroll/roll_site/conf/role_conf.json
@@ -1,7 +1,7 @@
 {
   "local": {
-    "role": "guest",
-    "party_id": 10001
+    "role": "host",
+    "party_id": 10002
   },
   "role": {
     "host": [
@@ -11,7 +11,7 @@
       99999
     ],
     "guest": [
-      10003
+      10001
     ]
   }
 }
