+++ python/eggroll/roll_site/conf/transfer_conf.json
@@ -1,7 +1,7 @@
 {
   "RsaIntersectTransferVariable": {
     "rsa_pubkey": {
-      "src": "guest",
+      "src": "host",
       "dst": [
         "host"
       ]
